LPS-54918 Copy code to portlet no logic changes
